[REEF-587] Create a default implementation for DriverRestartManager
This addressed the issue by
  * Changing DriverRestartManager from an interface to a final class.
  * Removing optional DriverRestartManager parameters.
  * Adding a default DriverRuntimeRestartManager that fails on any restart related options.

JIRA:
  [REEF-587](https://issues.apache.org/jira/browse/REEF-587)
