+++ lang/java/reef-common/src/main/java/org/apache/reef/runtime/common/driver/evaluator/EvaluatorManager.java
@@ -21,6 +21,7 @@ package org.apache.reef.runtime.common.driver.evaluator;
 import org.apache.reef.annotations.audience.DriverSide;
 import org.apache.reef.annotations.audience.Private;
 import org.apache.reef.driver.evaluator.CLRProcessFactory;
+import org.apache.reef.driver.parameters.EvaluatorConfigurationProviders;
 import org.apache.reef.driver.restart.DriverRestartManager;
 import org.apache.reef.driver.restart.DriverRestartUtilities;
 import org.apache.reef.tang.ConfigurationProvider;
@@ -29,7 +30,6 @@ import org.apache.reef.driver.context.FailedContext;
 import org.apache.reef.driver.evaluator.AllocatedEvaluator;
 import org.apache.reef.driver.evaluator.EvaluatorDescriptor;
 import org.apache.reef.driver.evaluator.JVMProcessFactory;
-import org.apache.reef.driver.parameters.EvaluatorConfigurationProviders;
 import org.apache.reef.driver.task.FailedTask;
 import org.apache.reef.exception.EvaluatorException;
 import org.apache.reef.exception.EvaluatorKilledByResourceManagerException;
@@ -104,7 +104,7 @@ public final class EvaluatorManager implements Identifiable, AutoCloseable {
   private final Set<ConfigurationProvider> evaluatorConfigurationProviders;
   private final JVMProcessFactory jvmProcessFactory;
   private final CLRProcessFactory clrProcessFactory;
-  private final Optional<DriverRestartManager> driverRestartManager;
+  private final DriverRestartManager driverRestartManager;
 
   // Mutable fields
   private Optional<TaskRepresenter> task = Optional.empty();
@@ -130,63 +130,10 @@ public final class EvaluatorManager implements Identifiable, AutoCloseable {
       final LoggingScopeFactory loggingScopeFactory,
       @Parameter(EvaluatorConfigurationProviders.class)
       final Set<ConfigurationProvider> evaluatorConfigurationProviders,
-      final JVMProcessFactory jvmProcessFactory,
-      final CLRProcessFactory clrProcessFactory,
-      final DriverRestartManager driverRestartManager) {
-    this(clock, remoteManager, resourceReleaseHandler, resourceLaunchHandler, evaluatorId, evaluatorDescriptor,
-        contextRepresenters, configurationSerializer, messageDispatcher, evaluatorControlHandler,
-        contextControlHandler, stateManager, exceptionCodec, idlenessSource, loggingScopeFactory,
-        evaluatorConfigurationProviders, jvmProcessFactory, clrProcessFactory, Optional.of(driverRestartManager));
-  }
-
-  @Inject
-  private EvaluatorManager(
-      final Clock clock,
-      final RemoteManager remoteManager,
-      final ResourceReleaseHandler resourceReleaseHandler,
-      final ResourceLaunchHandler resourceLaunchHandler,
-      @Parameter(EvaluatorIdentifier.class) final String evaluatorId,
-      @Parameter(EvaluatorDescriptorName.class) final EvaluatorDescriptorImpl evaluatorDescriptor,
-      final ContextRepresenters contextRepresenters,
-      final ConfigurationSerializer configurationSerializer,
-      final EvaluatorMessageDispatcher messageDispatcher,
-      final EvaluatorControlHandler evaluatorControlHandler,
-      final ContextControlHandler contextControlHandler,
-      final EvaluatorStatusManager stateManager,
-      final ExceptionCodec exceptionCodec,
-      final EventHandlerIdlenessSource idlenessSource,
-      final LoggingScopeFactory loggingScopeFactory,
-      @Parameter(EvaluatorConfigurationProviders.class)
-      final Set<ConfigurationProvider> evaluatorConfigurationProviders,
-      final JVMProcessFactory jvmProcessFactory,
-      final CLRProcessFactory clrProcessFactory) {
-    this(clock, remoteManager, resourceReleaseHandler, resourceLaunchHandler, evaluatorId, evaluatorDescriptor,
-        contextRepresenters, configurationSerializer, messageDispatcher, evaluatorControlHandler,
-        contextControlHandler, stateManager, exceptionCodec, idlenessSource, loggingScopeFactory,
-        evaluatorConfigurationProviders, jvmProcessFactory, clrProcessFactory, Optional.<DriverRestartManager>empty());
-  }
-
-  private EvaluatorManager(
-      final Clock clock,
-      final RemoteManager remoteManager,
-      final ResourceReleaseHandler resourceReleaseHandler,
-      final ResourceLaunchHandler resourceLaunchHandler,
-      final String evaluatorId,
-      final EvaluatorDescriptorImpl evaluatorDescriptor,
-      final ContextRepresenters contextRepresenters,
-      final ConfigurationSerializer configurationSerializer,
-      final EvaluatorMessageDispatcher messageDispatcher,
-      final EvaluatorControlHandler evaluatorControlHandler,
-      final ContextControlHandler contextControlHandler,
-      final EvaluatorStatusManager stateManager,
-      final ExceptionCodec exceptionCodec,
-      final EventHandlerIdlenessSource idlenessSource,
-      final LoggingScopeFactory loggingScopeFactory,
-      final Set<ConfigurationProvider> evaluatorConfigurationProviders,
       // TODO: Eventually remove the factories when they are removed from AllocatedEvaluatorImpl
       final JVMProcessFactory jvmProcessFactory,
       final CLRProcessFactory clrProcessFactory,
-      final Optional<DriverRestartManager> driverRestartManager) {
+      final DriverRestartManager driverRestartManager) {
     this.contextRepresenters = contextRepresenters;
     this.idlenessSource = idlenessSource;
     LOG.log(Level.FINEST, "Instantiating 'EvaluatorManager' for evaluator: {0}", evaluatorId);
@@ -396,7 +343,8 @@ public final class EvaluatorManager implements Identifiable, AutoCloseable {
         this.evaluatorControlHandler.setRemoteID(evaluatorRID);
         this.stateManager.setRunning();
 
-        boolean restartCompleted = this.driverRestartManager.get().evaluatorRecovered(this.evaluatorId);
+        final boolean restartCompleted =
+            this.driverRestartManager.onRecoverEvaluatorIsRestartComplete(this.evaluatorId);
 
         LOG.log(Level.FINE, "Received recovery heartbeat from evaluator {0}.", this.evaluatorId);
 
@@ -405,8 +353,8 @@ public final class EvaluatorManager implements Identifiable, AutoCloseable {
           LOG.log(Level.INFO, "All expected evaluators checked in.");
         } else {
           LOG.log(Level.INFO, "Expecting [{0}], [{1}] have checked in.",
-              new Object[]{this.driverRestartManager.get().getPreviousEvaluatorIds(),
-                  this.driverRestartManager.get().getRecoveredEvaluatorIds()});
+              new Object[]{this.driverRestartManager.getPreviousEvaluatorIds(),
+                  this.driverRestartManager.getRecoveredEvaluatorIds()});
         }
       }
